building front-end authentication pages and testing user endpoints with front-end
