+++ srcs/backend/transcendence/urls.py
@@ -5,7 +5,7 @@ from django.conf import settings
 from django.conf.urls.static import static
 
 urlpatterns = [
-	path('', TemplateView.as_view(template_name="index.html")),	# Home page
+	path('', TemplateView.as_view(template_name="auth.html")),	# Home page
     path('admin/', admin.site.urls),
 	path('user/', include('user.urls')),
 	path('game/', include('game.urls')),
