+++ srcs/backend/user/serializers.py
@@ -1,9 +1,7 @@
 import pyotp
 from .models import User, FriendRequest
 from transcendence import settings
-from django.utils.translation import gettext_lazy as _
 from django.core.mail import EmailMultiAlternatives
-from django.core.exceptions import ValidationError
 from django.contrib.auth import authenticate
 from django.template.loader import render_to_string
 from rest_framework import serializers
@@ -24,7 +22,7 @@ class UserCreateSerializer(serializers.ModelSerializer):
 	def create(self, validated_data):
 		
 		if User.objects.filter(email=validated_data['email'], is_active=True).exists():
-			raise serializers.ValidationError(_("Registration impossible, this email is already used by a active user."))
+			raise serializers.ValidationError("Registration impossible, this email is already used by a active user")
 		
 		password = validated_data.pop('password')
 		user = super().create(validated_data)
@@ -44,7 +42,7 @@ class UserCreateSerializer(serializers.ModelSerializer):
 			)
 		except Exception as e:
 			user.delete()
-			raise serializers.ValidationError(_(f"Error: send activation mail failed : {str(e)}"))
+			raise serializers.ValidationError(f"Error: send activation mail failed : {str(e)}")
 		
 		return user
 
@@ -74,15 +72,10 @@ class UserLoginSerializer(serializers.Serializer):
 		username = attrs.get('username')
 		password = attrs.get('password')
 		if not username or not password:
-			raise serializers.ValidationError(_("Username and password are required."))
+			raise serializers.ValidationError({"message": "Username and password are required"})
 		user = authenticate(username=username, password=password)
-		if user is None:
-			raise serializers.ValidationError(_("Non-active account or Invalid username/password."))
-		if user.is_staff:
-			raise serializers.ValidationError({
-			    "detail": "Redirect to admin login",
-			    "redirect_url": "/admin/login/"
-			})
+		if user is None or user.is_staff:
+			raise serializers.ValidationError({"message": "Non-active account or Invalid username/password"})
 		self._send_2FA_mail(user)
 		attrs['user'] = user
 		return attrs
@@ -93,10 +86,10 @@ class OtpCodeChecking(serializers.Serializer):
 	def validate_otp_code(self, value):
 		user = self.context.get('user')
 		if user is None:
-			raise serializers.ValidationError(_("User not found."))
+			raise serializers.ValidationError("User not found.")
 		totp = pyotp.TOTP(user.otp_secret, interval=300)
 		if not totp.verify(value, for_time=datetime.now(timezone.utc)):
-			raise serializers.ValidationError(_("Invalid OTP code."))
+			raise serializers.ValidationError("Invalid OTP code.")
 		return value
 
 class UserViewSetSerializer(serializers.ModelSerializer):
@@ -111,16 +104,16 @@ class UserViewSetSerializer(serializers.ModelSerializer):
 
 		if 'password' in validated_data:
 			if 'current_password' not in validated_data or validated_data['current_password'] is None:
-				raise serializers.ValidationError(_("The current password is required to update your password."))
+				raise serializers.ValidationError("The current password is required to update your password.")
 			elif not instance.check_password(validated_data['current_password']):
-				raise serializers.ValidationError(_("The current password does not match the existing password."))
+				raise serializers.ValidationError("The current password does not match the existing password.")
 			instance.set_password(validated_data['password'])
-			success_messages.append(_("Password changed successfully !"))
+			success_messages.append("Password changed successfully !")
 			validated_data.pop('password')
 
 		if 'new_email' in validated_data and validated_data['new_email'] is not None:
 			if User.objects.filter(email=validated_data['new_email'], is_active=True).exists():
-				raise serializers.ValidationError(_("Changing to new email impossible, this email is already used by a active user."))
+				raise serializers.ValidationError("Changing to new email impossible, this email is already used by a active user.")
 			try:
 				send_activation_email(
 					instance,
@@ -130,21 +123,21 @@ class UserViewSetSerializer(serializers.ModelSerializer):
 					'emails/mail_changed.txt',
 					'emails/mail_changed.html'
 					)
-				success_messages.append(_("Send new email confirmation successfully !"))
+				success_messages.append("Send new email confirmation successfully !")
 			except Exception as e:
-				raise serializers.ValidationError(_(f"Error: send mail to new email adress failed : {str(e)}"))
+				raise serializers.ValidationError(f"Error: send mail to new email adress failed : {str(e)}")
 		
 		# reste a tester avec le front-end
 		if 'friends' in validated_data:
 			new_friend = validated_data['friends']
 			if instance.friends.filter(username=new_friend.username).exists():
-				success_messages.append(_(f"{new_friend.username} is already in your actual friends list"))
+				success_messages.append(f"{new_friend.username} is already in your actual friends list")
 			else:
 				if not FriendRequest.objects.filter(sender=instance, receiver=new_friend).exists():
 					FriendRequest.objects.create(sender=instance, receiver=new_friend)
-					success_messages.append(_(f"Friend request sent to {new_friend.username} successfully."))
+					success_messages.append(f"Friend request sent to {new_friend.username} successfully.")
 				else:
-					success_messages.append(_(f"You have already sent a friend request to {new_friend.username}."))
+					success_messages.append(f"You have already sent a friend request to {new_friend.username}.")
 			validated_data.pop('friends')
 
 		for attr, value in validated_data.items():
