+++ srcs/backend/user/views.py
@@ -6,6 +6,7 @@ from django.utils.http import urlsafe_base64_decode
 from django.utils.translation import gettext_lazy as _
 from django.shortcuts import get_object_or_404
 from rest_framework import status
+from rest_framework.exceptions import ValidationError
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.decorators import api_view
@@ -66,39 +67,50 @@ class UserLoginView(APIView):
       
 	def post(self, request, *args, **kwargs):
 		serializer = self.serializer_class(data=request.data)
-		serializer.is_valid(raise_exception=True)
-
-		user = serializer.validated_data['user']
-		otp_verification_url = reverse('otp_verification', args=[user.id])
-
-		return Response({
-		    "message": "Send verification code successfully",
-		    "otp_verification_url": otp_verification_url
-		}, status=status.HTTP_200_OK)
-
+		try:
+			serializer.is_valid(raise_exception=True)
+			user = serializer.validated_data['user']
+			otp_verification_url = reverse('otp_verification', args=[user.id])
+			return Response({
+			    "message": "A verification code is sent to your email",
+			    "otp_verification_url": otp_verification_url
+			}, status=status.HTTP_200_OK)
+		
+		except ValidationError as e:
+			error_message = str(e.detail.get("otp_code", ["An error occurred"])[0])
+			print(f"M={error_message}")
+			return Response({"message": error_message}, status=status.HTTP_401_UNAUTHORIZED)
+		
 class OtpVerificationView(APIView):
+	model = User
 	permission_classes = [AllowAny]
 	serializer_class = OtpCodeChecking
 
 	def post(self, request, user_id, *args, **kwargs):
-		user = get_object_or_404(get_user_model(), id=user_id)
-
+		user = User.objects.get(id=user_id)
 		serializer = self.serializer_class(data=request.data, context={'user': user})
-		serializer.is_valid(raise_exception=True)
+		try:
+			serializer.is_valid(raise_exception=True)
 
-		refresh = RefreshToken.for_user(user)
-		access_token = str(refresh.access_token)
-		refresh_token = str(refresh)
+			refresh = RefreshToken.for_user(user)
+			access_token = str(refresh.access_token)
+			refresh_token = str(refresh)
 
-		response = Response({"access_token": access_token}, status=status.HTTP_200_OK)
-		cookie_max_age = 3600 * 2
-		response.set_cookie(
-			'refresh_token',
-		    refresh_token,
-		    max_age=cookie_max_age,
-		    httponly=True
-		)
-		return response
+			response = Response({"access_token": access_token}, status=status.HTTP_200_OK)
+			cookie_max_age = 3600 * 24
+			response.set_cookie(
+				'refresh_token',
+			    refresh_token,
+			    max_age=cookie_max_age,
+			    httponly=True,
+				secure=True
+			)
+			return response
+		except ValidationError as e:
+			error_message = e.detail.get("message")
+			# return Response({"message": error_message}, status=status.HTTP_401_UNAUTHORIZED)
+			return Response({"message": str(e)}, status=status.HTTP_401_UNAUTHORIZED)
+
 
 class LogoutView(APIView):
     permission_classes = [IsAuthenticated]
