+++ srcs/static/user/css/auth.css
@@ -40,7 +40,7 @@ body {
     padding-right: 20px;
 }
 
-.main-content .left .dynamic-box {
+.main-content .left .auth-box {
     border: 1px solid #000; /* Bordure de la boîte dynamique */
     padding: 20px;
 }
