+++ srcs/templates/auth.html
@@ -5,7 +5,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Transcendence Game</title>
-    <link rel="stylesheet" type="text/css" href="{% static 'global/css/app.css' %}?v=1.0">
+    <link rel="stylesheet" type="text/css" href="{% static 'user/css/auth.css' %}?v=1.0">
 </head>
 <body>
     <div class="game-frame">
@@ -14,11 +14,11 @@
                 <a href="#" class="home">Home</a>
                 <h3 class="project-name">Transcendence</h3>
             </div>
-            <button id="profileButton">P</button>
+            <button id="profileButton">Profile</button>
         </div>
         <div class="main-content">
             <div class="left">
-                <div class="dynamic-box" id="dynamicBox">
+                <div class="auth-box" id="authBox">
                     <!-- Le contenu dynamique sera chargé ici -->
                 </div>
             </div>
@@ -40,8 +40,7 @@
         <div id="responseData">
             <!-- Les données de connexion s'afficheront ici -->
         </div>
-        <div id="app"></div>
     </div>
-    <script type="module" src="{% static 'global/js/app.js' %}?v=1.1"></script>
+    <script type="module" src="{% static 'user/js/auth.js' %}?v=1.1"></script>
 </body>
 </html>
